Flytt lenker og knapper ut i egen komponent
