+++ src/frontend/Felles/Visittkort/Visittkort.tsx
@@ -1,39 +1,21 @@
 import React from 'react';
 import styles from './Visittkort.module.css';
-import {
-    Behandling,
-    Klagebehandlingsårsak,
-    klagebehandlingsårsakTilTekst,
-    PåklagetVedtakstype,
-} from '../../App/typer/fagsak';
-import { HenleggKnapp } from './HenleggKnapp';
-import { BodyShort, CopyButton, HStack, Label, Link, Stack } from '@navikt/ds-react';
+import { Behandling } from '../../App/typer/fagsak';
+import { BodyShort, CopyButton, HStack, Label } from '@navikt/ds-react';
 import { PersonStatusVarsel } from '../Varsel/PersonStatusVarsel';
 import { AdressebeskyttelseVarsel } from '../Varsel/AdressebeskyttelseVarsel';
-import { EtikettFokus, EtikettInfo, EtikettSuksess } from '../Varsel/Etikett';
+import { EtikettFokus } from '../Varsel/Etikett';
 import { erEtterDagensDato } from '../../App/utils/dato';
-import { stønadstypeTilTekst } from '../../App/typer/stønadstype';
-import { ExternalLinkIcon } from '@navikt/aksel-icons';
-import {
-    utledBehandlingLenke,
-    utledSaksoversiktLenke,
-    utledTilbakekrevingLenke,
-} from '../../App/utils/utils';
-import { useApp } from '../../App/context/AppContext';
-import { FagsystemType } from '../../Komponenter/Behandling/Formkrav/typer';
-import { SettPåVentKnapp } from './SettPåVentKnapp';
-import { EndreBehandlendeEnhetKnapp } from './EndreBehandlendeEnhetKnapp';
 import { IkonVelger } from '../IkonVelger/IkonVelger';
 import { formaterOrgNummer, Institusjon } from '../../App/typer/institusjon';
 import { usePersonopplysningerContext } from '../../App/context/PersonopplysningerContext';
+import { LenkerOgKnapper } from './LenkerOgKnapper';
 
 interface Props {
     behandling: Behandling;
 }
 
 export function Visittkort({ behandling }: Props) {
-    const { appEnv } = useApp();
-
     const {
         fagsakEier: {
             personIdent,
@@ -46,16 +28,6 @@ export function Visittkort({ behandling }: Props) {
             vergemål,
         },
     } = usePersonopplysningerContext();
-
-    const skalLenkeTilFagsystemBehandling =
-        behandling.påklagetVedtak.påklagetVedtakstype === PåklagetVedtakstype.VEDTAK &&
-        behandling.påklagetVedtak.fagsystemVedtak?.fagsystemType === FagsystemType.ORDNIÆR;
-    const skalLenkeTilTilbakekreving =
-        behandling.påklagetVedtak.påklagetVedtakstype === PåklagetVedtakstype.VEDTAK &&
-        behandling.påklagetVedtak.fagsystemVedtak?.fagsystemType === FagsystemType.TILBAKEKREVING;
-    const behandlingLenke = utledBehandlingLenke(behandling, appEnv.eksternlenker);
-    const saksoversiktLenke = utledSaksoversiktLenke(behandling, appEnv.eksternlenker);
-    const tilbakekrevingLenke = utledTilbakekrevingLenke(behandling, appEnv.eksternlenker);
 
     return (
         <div className={styles.container}>
@@ -103,48 +75,7 @@ export function Visittkort({ behandling }: Props) {
                         </div>
                     )}
                 </VisittkortInner>
-                <HStack align={'center'} justify={'start'} gap={'space-8'} padding={'space-8'}>
-                    {skalLenkeTilFagsystemBehandling && (
-                        <Link href={behandlingLenke} target="_blank">
-                            Gå til behandling
-                            <ExternalLinkIcon
-                                aria-label="Gå til behandling"
-                                fontSize={'1.375rem'}
-                            />
-                        </Link>
-                    )}
-                    {skalLenkeTilTilbakekreving && (
-                        <Link href={tilbakekrevingLenke} target="_blank">
-                            Gå til tilbakekreving
-                            <ExternalLinkIcon
-                                aria-label="Gå til tilbakekreving"
-                                fontSize={'1.375rem'}
-                            />
-                        </Link>
-                    )}
-                    <Link href={saksoversiktLenke} target="_blank">
-                        Gå til saksoversikt
-                        <ExternalLinkIcon aria-label="Gå til saksoversikt" fontSize={'1.375rem'} />
-                    </Link>
-                    {behandling && (
-                        <Stack direction={'row'} gap={'space-8'}>
-                            <HStack justify={'end'} gap={'space-8'}>
-                                <EtikettSuksess>
-                                    {stønadstypeTilTekst[behandling.stønadstype]}
-                                </EtikettSuksess>
-                                {behandling.årsak ===
-                                    Klagebehandlingsårsak.HENVENDELSE_FRA_KABAL && (
-                                    <EtikettInfo>
-                                        {klagebehandlingsårsakTilTekst[behandling.årsak]}
-                                    </EtikettInfo>
-                                )}
-                            </HStack>
-                            <SettPåVentKnapp />
-                            <EndreBehandlendeEnhetKnapp fagsystem={behandling.fagsystem} />
-                            <HenleggKnapp />
-                        </Stack>
-                    )}
-                </HStack>
+                <LenkerOgKnapper behandling={behandling} />
             </HStack>
         </div>
     );
